+++ tools/robocompdsl/CMakeLists.txt
@@ -7,6 +7,7 @@ cmake_minimum_required( VERSION 2.8)
 INSTALL(PROGRAMS "parseCDSL.py" DESTINATION /opt/robocomp/python)
 INSTALL(PROGRAMS "parseIDSL.py" DESTINATION /opt/robocomp/python)
 INSTALL(PROGRAMS "rcExceptions.py" DESTINATION /opt/robocomp/python)
+INSTALL(PROGRAMS "robocompdsl_core.py" DESTINATION /opt/robocomp/python)
 INSTALL(PROGRAMS "robocompdsl.py" DESTINATION /opt/robocomp/bin RENAME robocompdsl )
 
 INSTALL(DIRECTORY "templateCPP" DESTINATION /opt/robocomp/share/robocompdsl )
